Candidate fix for #47: Why did aggregate get a timestamp in March?
The logs show that time_offset had just been updated, which resulted in _last_nanode_time being set to 51753779 but a packet had arrived from channel 1 at 51753758 so the system believe that roll-over had occured so it added 2E32!  Woops!  So the fix is to not update _last_nanode_time in _set_time_offset()